When vsync is active FPS can't be limited
For what I undersand vsync syncronizes the application framerate to the monitor so limiting FPS when vsync is active doesn't make sense and doesn't work (we would still have to use SDL_Delay in combination with vsync I I don't see how it would make sense)

+++ Engine/Source/Timer.h
@@ -47,7 +47,7 @@ public:
     bool UpdateFpsLog() const { return mUpdateFpsLog; }
     void FpsLogUpdated() { mUpdateFpsLog = false; }
 
-    unsigned int GetLowestFPS() const { return mLowestFps; }
+    float GetLowestFPS() const { return mLowestFps; }
     long GetLowestFpsTime() const { return mLowestFpsTime; }
     void SetLowestFps();
 
@@ -90,7 +90,7 @@ private:
 
     bool mUpdateFpsLog = false;         //True if enough time has passed (500ms) to calculate FPS
 
-    unsigned int mLowestFps = mFpsLimit;
+    float mLowestFps = mFpsLimit;
     long mLowestFpsTime = 0;
 
     long mFrameDelay = 0;               //Time in ms the last frame was delayed for
